On refresh, use project's device list to find device
Rather than querying the API for the device ID. This allows handling of
errors where the device was deleted since the last refresh and assigned
to another Packet user.

Fix errors when termination_time is ""

"" should mean "argument not set" in the config. Fixes an issue when
termination_time is set to "" in a terraform module.